% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/motifEnrichment.R
\name{motifEnrichment}
\alias{motifEnrichment}
\title{Test for motif enrichment significance}
\usage{
motifEnrichment(target.promoters, universe.promoters, all.counts = F,
  motifs = motifsSS)
}
\arguments{
\item{target.promoters}{A dataframe with target promoters.}

\item{universe.promoters}{A dataframe with the total number of promotors to test against.}

\item{motifs}{An output from DNAStringSet()}
}
\value{
A data frame of tested motifs with sorted p-values
}
\description{
Test for motif enrichment significance
}
\examples{
motifEnrichment(my.targets, my.universe, all.counts = F, motifs=motifsSS)
}

